Refactor CLI help for 3ae2f31. Minor namespace alt
