+++ djs/__init__.py
@@ -1,3 +1,3 @@
-from .job_scheduler.import_helper import *
+from .job_scheduler.api import *
 from .perturbation_engine.from_yaml import perturb_from_yaml
 from .perturbation_engine.parameter_editor import perturb_parameters
